Dosbox svn 3715 - Add midnight/next day increment - Add date and time commands (functionality limited)
git-svn-id: https://svn.code.sf.net/p/jdosbox/code/trunk@597 4cb9e868-4783-4b93-94d1-9e8e689819c6

+++ src/jdos/dos/Dos_Date.java
@@ -1,7 +1,7 @@
 package jdos.dos;
 
 public class Dos_Date {
-    public /*Bit16u*/short year;
-	public /*Bit8u*/byte month;
-	public /*Bit8u*/byte day;
+    public /*Bit16u*/int year;
+	public /*Bit8u*/short month;
+	public /*Bit8u*/short day;
 }
